+++ src/exec/PageExecutor.ts
@@ -6,7 +6,7 @@ export class PageExecutor extends SimpleExecutor {
         return type === 'page';
     }
         
-    async execute(module: MiniModule, started: boolean): Promise<void> {
+    async execute(module: MiniModule): Promise<void> {
         for (const file of module.tmpFiles!) {
             // 将配置挂载到 pluto.assets[模块名]
             pluto.third.assets[module.name][file.name] = file.content;