+++ src/manager/ConfigManager.ts
@@ -3,42 +3,68 @@ import { ModStorage } from "storage";
 
 export class ConfigManager {
     
-    async getLive(name: string) {
+    async getLive(name: string, id: string) {
         const module = await ModStorage.loadModule(name);
         if(module.position === "Z") {
-            return this.getConfigOps(name);
+            return this.getConfigOps(name, id);
         }
-        return await this.getConfigLocal(name);
+        return await this.getConfigLocal(name, id);
     }
 
-    getCached(name: string) {
-        return this.getConfigOps(name);
+    getCached(name: string, id: string) {
+        return this.getConfigOps(name, id);
     }
 
-    getConfigOps(name: string) {
+    getConfigOps(name: string, id: string) {
+        if(id) return pluto.third.assets[name][`${id}.json`];
         const config = pluto.third.assets[name][name];
         if(config) return config;
         return pluto.third.assets[name]["data.json"];
     }
 
-    async getConfigLocal(name: string) {
-        const content = await app.vault.adapter.read(`${pluto.self.settings.configPath}/${name}.yaml`);
-        const config = parseYaml(content);
-        pluto.third.assets[name][name] = config;
-        if(config) return config;
+    async getConfigLocal(name: string, id: string) {
+        if(id) {
+            const config = await app.vault.readJson(`${pluto.self.settings.configPath}/${id}.json`);
+            pluto.third.assets[name][`${id}.json`] = config;
+            return config;
+        }else {
+            const content = await app.vault.readRaw(`${pluto.self.settings.configPath}/${name}.yaml`);
+            const config = parseYaml(content);
+            pluto.third.assets[name][name] = config;
+            if(config) return config;
+        }
     }
 
-    async write(name: string, list: any) {
+    async write(name: string, list: any, id: string) {
         const module = await ModStorage.loadModule(name);
         if(module.position === "Z") {
-            const config = module.files.find(f => f.name === "config.yaml");
-            if(config) {
-                config.content = stringifyYaml(list);
+            if(id) {
+                const config = module.files.find(f => f.name === `${id}.json`);
+                if(config) {
+                    config.content = JSON.stringify(list, void 0, 2);
+                }else {
+                    module.files.push({
+                        name: `${id}.json`,
+                        content: JSON.stringify(list, void 0, 2),
+                        type: "json",
+                    });
+                }
                 ModStorage.saveModule(module);
+            }else {
+                const config = module.files.find(f => f.name === "config.yaml");
+                if(config) {
+                    config.content = stringifyYaml(list);
+                    ModStorage.saveModule(module);
+                }
             }
         }else {
-            const yaml = stringifyYaml(list);
-            await app.vault.adapter.write(`${pluto.self.settings.configPath}/${name}.yaml`, yaml);
+            if(id) {
+                const content = JSON.stringify(list, void 0, 2);
+                await app.vault.adapter.write(`${pluto.self.settings.configPath}/${id}.json`, content);
+            }else {
+                const yaml = stringifyYaml(list);
+                await app.vault.adapter.write(`${pluto.self.settings.configPath}/${name}.yaml`, yaml);
+            }
         }
     }
 }