+++ src/manager/FormManager.ts
@@ -6,19 +6,20 @@ export class FormManager {
 
     async prompt(name: string, required: boolean = true, title: string = ''): Promise<string> {
         const config = FormJson.input(title, name, required);
-        const values = await this.openJson(config, {});
-        return values["name"] as string;
+        return this.openJson(config, {})
+            .then((values) => values["name"] as string);
     }
 
     async create(module: any): Promise<string> {
         const config = FormJson.create();
-        const values = await this.openJson(config, module);
-        return values["name"] as string;
+        return this.openJson(config, module)
+            .then((values) => values["name"] as string);
     }
 
     async openSetting(module: any): Promise<FormValues> {
         const config = FormJson.setting(module.name);
-        return await this.openJson(config, module as FormValues);
+        return this.openJson(config, module as FormValues)
+            .then((values) => values);
     }
 
     async openYaml(config: FormModalConfig, defaultValues: FormValues = {}): Promise<FormValues> {
@@ -27,7 +28,7 @@ export class FormManager {
                 console.log('表单提交值:', values);
                 resolve(values);
             }, () => {
-                reject({type: 'cancel', message: '用户取消了操作'});
+                reject({ type: 'cancel', message: '用户取消了操作' });
             }).open();
         });
     }
@@ -38,7 +39,7 @@ export class FormManager {
                 console.log('表单提交值:', values);
                 resolve(values);
             }, () => {
-                reject({type: 'cancel', message: '用户取消了操作'});
+                reject({ type: 'cancel', message: '用户取消了操作' });
             }).open();
         });
     }