+++ src/third/DvaComponent.ts
@@ -7,6 +7,6 @@ export class DvaComponent extends SimpleComponent {
 
     async execute(block: any): Promise<void> {
         super.execute(block);
-        this.api.renderValue(block.text, block.contentEl);
+        this.self.api.renderValue(block.text, block.contentEl);
     }
 }