+++ src/third/SimpleComponent.ts
@@ -2,14 +2,16 @@ import { Notice } from "obsidian";
 
 // 创建一个简单的ThirdComponent实现，用于其他插件
 export class SimpleComponent implements ThirdComponent {
-
-    op: any;
-    api: any;
-    codes: Map<string, any> = new Map();
+    
+    self: any;
     configPath: string;
+    codes: Map<string, any[]> = new Map();
 
     constructor(configPath: string) {
         this.configPath = configPath;
+    }
+    get prop(): PlutoProps {
+        throw new Error("Method not implemented.");
     }
 
     get pluginId(): string {
@@ -21,20 +23,24 @@ export class SimpleComponent implements ThirdComponent {
     }
 
     check(): void {
-        if (!this.op) throw new Error('Please install plugin ' + this.pluginId);
+        if (!pluto.third.react) throw new Error('Please install plugin ' + this.pluginId);
     }
     
     bind(op: any, prop: PlutoProps): ThirdComponent {
-        this.op = op;
-        this.api = op.api;
-        this.patch();
+        this.self = op;
         pluto.third[prop] = this;
+        this.patch();
         console.log(`[Pluto Hub] ${prop} successfully bound to pluto.third.${prop}`);
         return this;
     }
 
-    register(key: string, code: any) {
-        this.codes.set(key, code);
+    register(name: string, block: any) {
+        const current = this.codes.get(name);
+        if(current) {
+            current.push(block);
+        } else {
+            this.codes.set(name, [block]);
+        }
     }
 
     async load(params: ModParams): Promise<void> { }
@@ -44,8 +50,10 @@ export class SimpleComponent implements ThirdComponent {
     }
 
     async executeAll(): Promise<void> {
-        for (const block of this.codes.values()) {
-            await this.execute(block);
+        for (const blocks of this.codes.values()) {
+            for (const block of blocks) {
+                await this.execute(block);
+            }
         }
     }
 }